edge_ai/requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
tflite-runtime>=2.14.0
paho-mqtt>=1.6.1
RPi.GPIO>=0.7.1
